remove backwards compatibility * move SettingsPanelItem to own file
