+++ src/ts/components/settingspanel.ts
@@ -1,15 +1,10 @@
-import {ContainerConfig, Container} from './container';
-import {ComponentConfig, Component} from './component';
+import {Container, ContainerConfig} from './container';
 import {SelectBox} from './selectbox';
-import {Label} from './label';
 import {UIInstanceManager} from '../uimanager';
-import {VideoQualitySelectBox} from './videoqualityselectbox';
-import {AudioQualitySelectBox} from './audioqualityselectbox';
 import {Timeout} from '../timeout';
 import {Event, EventDispatcher, NoArgs} from '../eventdispatcher';
-import {ListBox} from './listbox';
-import {PlaybackSpeedSelectBox} from './playbackspeedselectbox';
 import {SettingsPanelPage} from './settingspanelpage';
+import {SettingsPanelItem} from './settingspanelitem';
 
 /**
  * Configuration interface for a {@link SettingsPanel}.
@@ -30,7 +25,7 @@ export interface SettingsPanelConfig extends ContainerConfig {
 }
 
 /**
- * A panel containing a list of {@link SettingsPanelItem items} that represent labelled settings.
+ * A panel containing a list of {@link SettingsPanelPage items}.
  */
 export class SettingsPanel extends Container<SettingsPanelConfig> {
 
@@ -40,10 +35,6 @@ export class SettingsPanel extends Container<SettingsPanelConfig> {
   private activePageIndex = 0;
   private navigationStack: SettingsPanelPage[] = [];
 
-  /**
-   * Needed for backwards compatibility
-   * @deprecated
-   */
   private settingsPanelEvents = {
     onSettingsStateChanged: new EventDispatcher<SettingsPanel, NoArgs>(),
   };
@@ -52,21 +43,6 @@ export class SettingsPanel extends Container<SettingsPanelConfig> {
 
   constructor(config: SettingsPanelConfig) {
     super(config);
-
-    // Add backwards compatibility
-    // When all components are a type of SettingsPanelItem create a single SettingsPanelPage
-    // and work with this newly created page
-    const isTypeOfSettingsPanelItem = (currentValue: Component<ComponentConfig>) => {
-      return currentValue instanceof SettingsPanelItem;
-    };
-
-    if (this.getComponents().every(isTypeOfSettingsPanelItem)) {
-      let mainPage = new SettingsPanelPage({
-        components: this.getComponents(),
-      });
-      this.config.components = [mainPage];
-      config.components = this.config.components;
-    }
 
     this.config = this.mergeConfig<SettingsPanelConfig>(config, {
       cssClass: 'ui-settings-panel',
@@ -108,7 +84,7 @@ export class SettingsPanel extends Container<SettingsPanelConfig> {
       });
     }
 
-    // pass event from root page through for backwards compatibility
+    // pass event from root page through
     this.getRootPage().onSettingsStateChanged.subscribe(() => {
       this.onSettingsStateChangedEvent();
     });
@@ -263,11 +239,11 @@ export class SettingsPanel extends Container<SettingsPanelConfig> {
   }
 
   /**
-   * Checks if there are active settings within this settings panel. An active setting is a setting that is visible
-   * and enabled, which the user can interact with.
+   * Checks if there are active settings within the root page of the settings panel.
+   * An active setting is a setting that is visible and enabled, which the user can interact with.
    * @returns {boolean} true if there are active settings, false if the panel is functionally empty to a user
    */
-  hasActiveSettings(): boolean {
+  rootPageHasActiveSettings(): boolean {
     return this.getRootPage().hasActiveSettings();
   }
 
@@ -292,118 +268,7 @@ export class SettingsPanel extends Container<SettingsPanelConfig> {
     this.settingsPanelEvents.onSettingsStateChanged.dispatch(this);
   }
 
-  /**
-   * Gets the event that is fired when one or more {@link SettingsPanelItem items} have changed state.
-   * @deprecated
-   * @returns {Event<SettingsPanel, NoArgs>}
-   */
   get onSettingsStateChanged(): Event<SettingsPanel, NoArgs> {
     return this.settingsPanelEvents.onSettingsStateChanged.getEvent();
-  }
-
-  addComponent(component: Component<ComponentConfig>) {
-    if (component instanceof SettingsPanelPage) {
-      super.addComponent(component);
-    } else {
-      // backwards compatibility
-      this.getRootPage().addComponent(component);
-    }
-  }
-}
-
-/**
- * An item for a {@link SettingsPanel},
- * Containing an optionnal {@link Label} and a component that configures a setting.
- * If the components is a {@link SelectBox} it will handle the logic of displaying it or not
- */
-export class SettingsPanelItem extends Container<ContainerConfig> {
-
-  private label: Component<ComponentConfig>;
-  private setting: Component<ComponentConfig>;
-
-  private settingsPanelItemEvents = {
-    onActiveChanged: new EventDispatcher<SettingsPanelItem, NoArgs>(),
-  };
-
-  constructor(label: string | Component<ComponentConfig>, setting: Component<ComponentConfig>, config: ContainerConfig = {}) {
-    super(config);
-
-    this.setting = setting;
-
-    this.config = this.mergeConfig(config, {
-      cssClass: 'ui-settings-panel-item',
-    }, this.config);
-
-    if (label !== null) {
-      if (label instanceof Component) {
-        this.label = label;
-      } else {
-        this.label = new Label({ text: label });
-      }
-      this.addComponent(this.label);
-    }
-
-    this.addComponent(this.setting);
-  }
-
-  configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager): void {
-    if (this.setting instanceof SelectBox || this.setting instanceof ListBox) {
-      let handleConfigItemChanged = () => {
-        if (!(this.setting instanceof SelectBox) && !(this.setting instanceof ListBox)) {
-          return;
-        }
-        // The minimum number of items that must be available for the setting to be displayed
-        // By default, at least two items must be available, else a selection is not possible
-        let minItemsToDisplay = 2;
-        // Audio/video quality select boxes contain an additional 'auto' mode, which in combination with a single
-        // available quality also does not make sense
-        if ((this.setting instanceof VideoQualitySelectBox && this.setting.hasAutoItem())
-          || this.setting instanceof AudioQualitySelectBox) {
-          minItemsToDisplay = 3;
-        }
-
-        if (this.setting.itemCount() < minItemsToDisplay) {
-          // Hide the setting if no meaningful choice is available
-          this.hide();
-        } else if (this.setting instanceof PlaybackSpeedSelectBox
-          && !uimanager.getConfig().playbackSpeedSelectionEnabled) {
-          // Hide the PlaybackSpeedSelectBox if disabled in config
-          this.hide();
-        } else {
-          this.show();
-        }
-
-        // Visibility might have changed and therefore the active state might have changed so we fire the event
-        // TODO fire only when state has really changed (e.g. check if visibility has really changed)
-        this.onActiveChangedEvent();
-      };
-
-      this.setting.onItemAdded.subscribe(handleConfigItemChanged);
-      this.setting.onItemRemoved.subscribe(handleConfigItemChanged);
-
-      // Initialize hidden state
-      handleConfigItemChanged();
-    }
-  }
-
-  /**
-   * Checks if this settings panel item is active, i.e. visible and enabled and a user can interact with it.
-   * @returns {boolean} true if the panel is active, else false
-   */
-  isActive(): boolean {
-    return this.isShown();
-  }
-
-  protected onActiveChangedEvent() {
-    this.settingsPanelItemEvents.onActiveChanged.dispatch(this);
-  }
-
-  /**
-   * Gets the event that is fired when the 'active' state of this item changes.
-   * @see #isActive
-   * @returns {Event<SettingsPanelItem, NoArgs>}
-   */
-  get onActiveChanged(): Event<SettingsPanelItem, NoArgs> {
-    return this.settingsPanelItemEvents.onActiveChanged.getEvent();
   }
 }
